+++ ocfl-java-itest/src/test/java/io/ocfl/itest/MutableHeadITest.java
@@ -29,6 +29,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
+import java.util.concurrent.Phaser;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
 import org.hamcrest.Matchers;
@@ -447,23 +448,25 @@ public abstract class MutableHeadITest {
             updater.writeFile(streamString("file2"), "file2.txt");
         });
 
+        var phaser = new Phaser(2);
+
         var future = CompletableFuture.runAsync(() -> {
             repo.stageChanges(ObjectVersionId.head(objectId), defaultVersionInfo, updater -> {
-                try {
-                    TimeUnit.SECONDS.sleep(3);
-                } catch (InterruptedException e) {
-                    throw new RuntimeException(e);
-                }
+                phaser.arriveAndAwaitAdvance();
+                phaser.arriveAndAwaitAdvance();
                 updater.writeFile(streamString("file3"), "file3.txt");
             });
         });
 
-        TimeUnit.MILLISECONDS.sleep(100);
+        phaser.arriveAndAwaitAdvance();
 
         repo.rollbackToVersion(ObjectVersionId.version(objectId, "v1"));
         repo.stageChanges(ObjectVersionId.head(objectId), defaultVersionInfo, updater -> {
             updater.writeFile(streamString("file4"), "file4.txt");
         });
+
+        phaser.arriveAndAwaitAdvance();
+        TimeUnit.MILLISECONDS.sleep(100);
 
         OcflAsserts.assertThrowsWithMessage(
                 ObjectOutOfSyncException.class,
@@ -497,23 +500,25 @@ public abstract class MutableHeadITest {
             updater.writeFile(streamString("file2"), "file2.txt");
         });
 
+        var phaser = new Phaser(2);
+
         var future = CompletableFuture.runAsync(() -> {
             repo.stageChanges(ObjectVersionId.head(objectId), defaultVersionInfo, updater -> {
-                try {
-                    TimeUnit.SECONDS.sleep(3);
-                } catch (InterruptedException e) {
-                    throw new RuntimeException(e);
-                }
+                phaser.arriveAndAwaitAdvance();
+                phaser.arriveAndAwaitAdvance();
                 updater.writeFile(streamString("file3"), "file3.txt");
             });
         });
 
-        TimeUnit.MILLISECONDS.sleep(100);
+        phaser.arriveAndAwaitAdvance();
 
         repo.rollbackToVersion(ObjectVersionId.version(objectId, "v1"));
         repo.updateObject(ObjectVersionId.head(objectId), defaultVersionInfo, updater -> {
             updater.writeFile(streamString("file4"), "file4.txt");
         });
+
+        phaser.arriveAndAwaitAdvance();
+        TimeUnit.MILLISECONDS.sleep(100);
 
         OcflAsserts.assertThrowsWithMessage(
                 ObjectOutOfSyncException.class,